refactor: separate usePlayers and useSocket
